backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.52.2
qdrant-client==1.12.1
langchain==0.3.7
langchain-openai==0.2.7
langchain-anthropic==0.2.7
langchain-community==0.3.7
langchain-qdrant==0.1.3
tiktoken==0.8.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlmodel==0.0.22
python-multipart==0.0.20